+++ src/Rocks.Analysis.IntegrationTests/RockContextTests.cs
@@ -24,7 +24,7 @@ public static class RockContextTests
 		}, Throws.TypeOf<ExpectationException>());
 
 	[Test]
-	public static void CausesVerificationExceptionAndContextIsEnabled() =>
+	public static void CausesVerificationException() =>
 		Assert.That(() =>
 		{
 			using var context = new RockContext();
@@ -37,20 +37,7 @@ public static class RockContextTests
 		}, Throws.TypeOf<VerificationException>());
 
 	[Test]
-	public static void CausesVerificationExceptionAndContextIsDisabled() =>
-		Assert.That(() =>
-		{
-			using var context = new RockContext(DisableVerification.Yes);
-			var expectations = context.Create<IContexturalCreateExpectations>();
-			expectations.Methods.DoStuff();
-			expectations.Methods.DoOtherStuff();
-
-			var mock = expectations.Instance();
-			mock.DoStuff();
-		}, Throws.Nothing);
-
-	[Test]
-	public static void CallbackCausesExceptionAndContextIsEnabled() =>
+	public static void CallbackCausesException() =>
 		Assert.That(() =>
 		{
 			using var context = new RockContext();
@@ -63,20 +50,7 @@ public static class RockContextTests
 		}, Throws.TypeOf<NotImplementedException>());
 
 	[Test]
-	public static void CallbackCausesExceptionAndContextIsDisabled() =>
-		Assert.That(() =>
-		{
-			using var context = new RockContext(DisableVerification.Yes);
-			var expectations = context.Create<IContexturalCreateExpectations>();
-			expectations.Methods.DoStuff()
-				.Callback(() => throw new NotImplementedException());
-
-			var mock = expectations.Instance();
-			mock.DoStuff();
-		}, Throws.TypeOf<NotImplementedException>());
-
-	[Test]
-	public static void CallbackCausesExceptionAndVerificationFailsAndContextIsEnabled() =>
+	public static void CallbackCausesExceptionAndVerificationFails() =>
 		Assert.That(() =>
 		{
 			using var context = new RockContext();
@@ -88,18 +62,4 @@ public static class RockContextTests
 			var mock = expectations.Instance();
 			mock.DoStuff();
 		}, Throws.TypeOf<VerificationException>());
-
-	[Test]
-	public static void CallbackCausesExceptionAndVerificationFailsAndContextIsDisabled() =>
-		Assert.That(() =>
-		{
-			using var context = new RockContext(DisableVerification.Yes);
-			var expectations = context.Create<IContexturalCreateExpectations>();
-			expectations.Methods.DoStuff()
-				.Callback(() => throw new NotImplementedException());
-			expectations.Methods.DoOtherStuff();
-
-			var mock = expectations.Instance();
-			mock.DoStuff();
-		}, Throws.TypeOf<NotImplementedException>());
 }